tjsp-scraper/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
asyncpg==0.29.0
bcrypt==4.2.0
beautifulsoup4==4.13.4
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==45.0.3
ecdsa==0.19.0
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
pdfminer.six==20250506
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
